storage/impl: Make SQL connection pooling functionality reusable via AbstractSQLConnectionPool
